+++ ExaminationSystem_API/ExaminationSystem_API/Dto/BranchDTO/BranchLookupDTO.cs
@@ -2,7 +2,7 @@ namespace ExaminationSystem_API.Dto.BranchDTO
 {
     public class BranchLookupDTO
     {
-        public int BranchId { get; set; }
+        public byte BranchId { get; set; }
         public string BranchName { get; set; }
     }
 }
