+++ ExaminationSystem_API/ExaminationSystem_API/Repository/ClassRepository/TrackRepository.cs
@@ -7,7 +7,7 @@ namespace ExaminationSystem_API.Repository.ClassRepository
         {
             _context = context;
         }
-        public async Task AddTrackWithStoredAsync(string name , int deptId) => await _context.Database.ExecuteSqlInterpolatedAsync($"EXEC [TrainingMangerStp].stp_AddTrack @TrackName = {name} , @DeptId = {deptId} ");
+        public async Task AddTrackWithStoredAsync(string name , byte deptId) => await _context.Database.ExecuteSqlInterpolatedAsync($"EXEC [TrainingMangerStp].stp_AddTrack @TrackName = {name} , @DeptId = {deptId} ");
         public async Task UpdateTrackWithStoredAsync(short trackId,string name , byte deptId) => await _context.Database.ExecuteSqlInterpolatedAsync($"EXEC [TrainingMangerStp].stp_UpdateTrack @TrackId = {trackId} , @TrackName = {name} , @DeptId = {deptId} ");
 
     }
