+++ ExaminationSystem_API/ExaminationSystem_API/Repository/InterfaceRepository/IBranchRepository.cs
@@ -5,9 +5,9 @@ namespace ExaminationSystem_API.Repository.InterfaceRepository
     public interface IBranchRepository : IGenericRepository<Branch>
     {
         Task AddBranchWithStoredAsync(string name);
-        Task UpdateBranchWithStoredAsync(int id, string name);
-        Task DeleteBranchWithStoredAsync(int id);
-        Task ActivateBranchWithStoredAsync(int id);
+        Task UpdateBranchWithStoredAsync(byte id, string name);
+        Task DeleteBranchWithStoredAsync(byte id);
+        Task ActivateBranchWithStoredAsync(byte id);
         Task<IEnumerable<VBranchsummary>> GetAllBranchSummaryWithStoredAsync();
     }
 }
