+++ ExaminationSystem_API/ExaminationSystem_API/Repository/InterfaceRepository/IDepartmentRepository.cs
@@ -2,8 +2,8 @@ namespace ExaminationSystem_API.Repository.InterfaceRepository
 {
     public interface IDepartmentRepository :IGenericRepository<Department>
     {
-        Task AddDepartmentWithStoredAsync(string name, int BranchID);
-        Task UpdateDepartmentWithStoredAsync(int DeptId, string name, int BranchID);
-        Task DeleteDepartmentWithStoredAsync(int DeptId);
+        Task AddDepartmentWithStoredAsync(string name, byte BranchID);
+        Task UpdateDepartmentWithStoredAsync(byte DeptId, string name, byte BranchID);
+        Task DeleteDepartmentWithStoredAsync(byte DeptId);
     }
 }
