+++ ExaminationSystem_API/ExaminationSystem_API/Service/InterfaceService/IBranchService.cs
@@ -7,9 +7,9 @@ namespace ExaminationSystem_API.Service.InterfaceService
     {
         Task<IEnumerable<BranchLookupDTO>> GetBranchesLookupAsync();
         Task AddBranchAsync(AddBranchDTO branchdto);
-        Task UpdateBranchAsync(int id, UpdateBranchDTO updateBranch);
-        Task DeleteBranchAsync(int id);
-        Task ActivateBranchAsync(int id);
+        Task UpdateBranchAsync(byte id, UpdateBranchDTO updateBranch);
+        Task DeleteBranchAsync(byte id);
+        Task ActivateBranchAsync(byte id);
         Task<PaginatedList<BranchSummaryDTO>> GetAllBranchSummryAsync(string? searchTerm, int pageNumber, int pageSize);
     }
 }
